# [slice]Optimizing bool index getitem paths