--- a/paddle/phi/infermeta/binary.cc
+++ b/paddle/phi/infermeta/binary.cc
@@ -2592,6 +2592,7 @@ void IndexElementwisePutWithTensorInferMeta(
 }
 
 void IndexElementwiseGetInferMeta(const MetaTensor& x,
+                                  const MetaTensor& gather_index,
                                   const std::vector<const MetaTensor*>& index,
                                   const std::vector<int64_t>& input_dims,
                                   const std::vector<int64_t>& input_strides,
@@ -2599,6 +2600,7 @@ void IndexElementwiseGetInferMeta(const MetaTensor& x,
                                   const std::vector<int64_t>& index_stride,
                                   const int64_t slice_offset,
                                   const bool accumulate,
+                                  const bool is_gather,
                                   MetaTensor* out) {
   out->set_dims(common::make_ddim(input_dims));
   out->set_dtype(x.dtype());