--- a/paddle/phi/infermeta/binary.h
+++ b/paddle/phi/infermeta/binary.h
@@ -486,6 +486,7 @@ void IndexElementwisePutWithTensorInferMeta(
     MetaTensor* out);
 
 void IndexElementwiseGetInferMeta(const MetaTensor& x,
+                                  const MetaTensor& gather_index,
                                   const std::vector<const MetaTensor*>& index,
                                   const std::vector<int64_t>& input_dims,
                                   const std::vector<int64_t>& input_strides,
@@ -493,6 +494,7 @@ void IndexElementwiseGetInferMeta(const MetaTensor& x,
                                   const std::vector<int64_t>& index_stride,
                                   const int64_t slice_offset,
                                   const bool accumulate,
+                                  const bool is_gather,
                                   MetaTensor* out);
 
 void KronInferMeta(const MetaTensor& x, const MetaTensor& y, MetaTensor* out);