--- a/paddle/phi/kernels/cpu/index_elementwise_get_kernel.cc
+++ b/paddle/phi/kernels/cpu/index_elementwise_get_kernel.cc
@@ -93,6 +93,7 @@ void CPUIndexElementwiseGetKernel(const phi::CPUContext& dev_ctx,
 template <typename T, typename Context>
 void IndexElementwiseGetKernel(const Context& dev_ctx,
                                const DenseTensor& x,
+                               const DenseTensor& gather_index,
                                const std::vector<const DenseTensor*>& index,
                                const std::vector<int64_t>& input_dims,
                                const std::vector<int64_t>& input_strides,
@@ -100,6 +101,7 @@ void IndexElementwiseGetKernel(const Context& dev_ctx,
                                const std::vector<int64_t>& index_stride,
                                const int64_t slice_offset,
                                const bool accumulate,
+                               const bool is_gather,
                                DenseTensor* out) {
   const auto& index_type = index[0]->dtype();
   PADDLE_ENFORCE_EQ(index_type == phi::DataType::INT64,