--- a/paddle/phi/kernels/gpu/index_elementwise_get_kernel.cu
+++ b/paddle/phi/kernels/gpu/index_elementwise_get_kernel.cu
@@ -19,6 +19,7 @@
 #include "paddle/phi/core/kernel_registry.h"
 #include "paddle/phi/kernels/funcs/index_elementwise.cu.h"
 #include "paddle/phi/kernels/funcs/stride_utils.h"
+#include "paddle/phi/kernels/gather_kernel.h"
 
 namespace phi {
 template <typename T, typename IndexT = int>
@@ -110,6 +111,7 @@ void GPUIndexElementwiseGetKernel(const phi::GPUContext& dev_ctx,
 template <typename T, typename Context>
 void IndexElementwiseGetKernel(const Context& dev_ctx,
                                const DenseTensor& x,
+                               const DenseTensor& gather_index,
                                const std::vector<const DenseTensor*>& index,
                                const std::vector<int64_t>& input_dims,
                                const std::vector<int64_t>& input_strides,
@@ -117,16 +119,8 @@ void IndexElementwiseGetKernel(const Context& dev_ctx,
                                const std::vector<int64_t>& index_stride,
                                const int64_t slice_offset,
                                const bool accumulate,
+                               const bool is_gather,
                                DenseTensor* out) {
-  const auto& index_type = index[0]->dtype();
-  PADDLE_ENFORCE_EQ(index_type == phi::DataType::INT64,
-                    true,
-                    common::errors::InvalidArgument(
-                        "Index holds the wrong type, it holds [%s], but "
-                        "desires to be [%s].",
-                        index_type,
-                        phi::DataType::INT64));
-
   auto out_dims = out->dims();
   if (out_dims.size() > 0) {
     std::vector<int64_t> output_dims(input_dims);
@@ -135,6 +129,20 @@ void IndexElementwiseGetKernel(const Context& dev_ctx,
 
   dev_ctx.template Alloc<T>(out);
   if (out->numel() == 0) return;
+
+  if (is_gather) {
+    phi::GatherKernel<T, Context>(dev_ctx, x, gather_index, 0, out);
+    return;
+  }
+
+  const auto& index_type = index[0]->dtype();
+  PADDLE_ENFORCE_EQ(index_type == phi::DataType::INT64,
+                    true,
+                    common::errors::InvalidArgument(
+                        "Index holds the wrong type, it holds [%s], but "
+                        "desires to be [%s].",
+                        index_type,
+                        phi::DataType::INT64));
 
   GPUIndexElementwiseGetKernel<T, int64_t>(dev_ctx,
                                            x,