--- a/paddle/phi/kernels/index_elementwise_get_grad_kernel.h
+++ b/paddle/phi/kernels/index_elementwise_get_grad_kernel.h
@@ -22,6 +22,7 @@ namespace phi {
 template <typename T, typename Context>
 void IndexElementwiseGetGradKernel(const Context& ctx,
                                    const DenseTensor& x,
+                                   const DenseTensor& gather_index,
                                    const std::vector<const DenseTensor*>& index,
                                    const DenseTensor& out_grad,
                                    const std::vector<int64_t>& input_dims,
@@ -30,6 +31,7 @@ void IndexElementwiseGetGradKernel(const Context& ctx,
                                    const std::vector<int64_t>& index_strides,
                                    const int64_t slice_offset,
                                    const bool accumulate,
+                                   const bool is_gather,
                                    DenseTensor* x_grad);
 
 }  // namespace phi