### PR Category
Performance Optimization

### PR Types
Others

### Description
bool_index.shape().size() == 1场景前向使用gather kernel，反向使用index_elementwise_get_grad kernel可以获得更好的性能。
例如： x.shape =  (108, 64, 12288) , index =  True
前向：807.00 us(index_elementwise_get) -> 518.00 us(gather)
反向：1611.02 us(gather grad) -> 1332.00 us(index_elementwise_get_grad)
pcard-67164